Spin off retryability to a separate module.
