+++ src/installfedoraonzfs/vm.py
@@ -12,13 +12,11 @@ import time
 
 from installfedoraonzfs.breakingbefore import BreakingBefore
 from installfedoraonzfs.cmd import Popen
+from installfedoraonzfs.retry import Retryable
 
 
 logger = logging.getLogger("VM")
 qemu_full_emulation_factor = 20
-
-
-class Retryable: pass
 
 
 class BootloaderWedged(Exception): pass
@@ -28,36 +26,6 @@ class SystemdSegfault(Retryable, Exception): pass
 
 
 class MachineNeverShutoff(Exception): pass
-
-
-class retry(object):
-    """Returns a callable that will retry the callee up to N times,
-    if the callee raises an exception of type Retryable.
-    To be clear: if N == 0, then the function will not retry.
-    So, to get three tries, you must pass N == 2."""
-
-    def __init__(self, N):
-        self.N = N
-
-    def __call__(self, kallable):
-
-        def retryer(*a, **kw):
-            logger = logging.getLogger("retry")
-            while True:
-                try:
-                    return kallable(*a, **kw)
-                except Retryable:
-                    if self.N >= 1:
-                        logger.error(
-                            "Received retryable error running %s, trying %s more times",
-                            kallable,
-                            self.N
-                        )
-                    else:
-                        raise
-                self.N -= 1
-
-        return retryer
 
 
 class Babysitter(threading.Thread):
